重构：将chunkType由uint改为AuraChunkType枚举
统一资源与节点类型标识为AuraChunkType枚举，重构相关特性、序列化、注册表及收集器接口，提升类型安全性和可读性，减少魔法数字，便于后续维护和扩展。保留uint构造重载以兼容旧代码。

+++ src/Aura3D.Core/Serialization/AuraFileWriter.cs
@@ -40,7 +40,7 @@ public class AuraFileWriter
         writer.Write(AuraFileHeader.Magic[3]);
         writer.Write(AuraFileHeader.CurrentFileVersion);
         writer.Write((uint)stringTableStream.Length);
-        writer.Write(_collector.RootChunkType);
+        writer.Write((uint)_collector.RootChunkType);
         writer.Write(_collector.RootResourceId);
 
         stringTableStream.Position = 0;
@@ -55,7 +55,7 @@ public class AuraFileWriter
         if (resource is not IAuraSerializable serializable)
             throw new InvalidOperationException($"Resource {resource.GetType().Name} does not implement IAuraSerializable.");
 
-        writer.Write(AuraResourceTypeRegistry.GetChunkType(resource));
+        writer.Write((uint)AuraResourceTypeRegistry.GetChunkType(resource));
         writer.Write(AuraResourceTypeRegistry.GetChunkVersion(resource));
 
         var stream = writer.BaseStream;
